omnichannel-showcase-democlient/code/sms-connector/requirements.txt
Flask==2.1.0
requests==2.26.0
requests-oauthlib==1.3.0
azure-communication-sms==1.0.1
